Check for std::future
std::future is not available for all architectures, e.g. it is missing
for ARMv5 (soft-float).

The problem is that libstdc++ only enables `std::future` if
`ATOMIC_INT_LOCK_FREE > 1` which is not true for the ARMv5 target.

As the future functionality is not used for much we just ifdef out the
parts from `std::future`.

Closes: #8