linnsongcast: allow setting receivers to play without looking for the sender
If a multi-room group has been established and all devices are turned
off and later on again, the receivers might be up and ready before the
sender is. In this case, the receivers will not find the sender and no
link is established between these devices.

That's why it is quite handy to allow setting up the receivers without
actually searching for the sender. Of course, this is only useful if a
sender has aready been set to the Songcast receiver service, e.g. by
getting the sender uri and metadata from a config state file.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>