Make additional escaping of URLs from Content Directories optional and non-default
Performing additional escaping of, e.g., single quote characters is useful with some Media
Servers and Renderers combinations, but it also causes issues with other renderers, e.g. Yamahas,
which emit currentTrackURI change events with a (maybe partially) decoded URL, which breaks string
comparisons in Control Points. This is not nice as neither approach will work in all cases, and the
problem can't be handled purely in the CP using the lib as there maybe a proxy CP involved (e.g,
Bubble server, in which the same issues are triggered by the differing URL).